MGDCTRS-1667: feat: init
