+++ package.json
@@ -9,7 +9,7 @@
     "@patternfly/quickstarts": "1.2.2",
     "@patternfly/react-catalog-view-extension": "4.92.26",
     "@patternfly/react-charts": "6.88.7",
-    "@patternfly/react-core": "4.235.7",
+    "@patternfly/react-core": "4.250.1",
     "@patternfly/react-icons": "4.86.7",
     "@patternfly/react-styles": "4.11.16",
     "@patternfly/react-table": "4.104.7",
